requirements.txt
Flask==1.0.2
Flask-Sockets==0.2.1
gunicorn==19.9.0
requests==2.21.0
google-cloud-speech
google-cloud-texttospeech

